KmsMediaHttpPostEndPoint: Raise EOS event
Application layer has no way to know when a media file has been
completely processed, this is the reason why we need to notify when
an HttpPostEndPoint has reached the end of a file. Application layer,
upon receiving this event could stop the recorder end point,
unregister any associated HttpPostEndPoint or even release the
whole pipeline if it was needed.

Change-Id: I18bd417e23b53936c00c3563b64d9f4531c9a540